+++ src/main/java/edu/kis/vh/nursery/FIFORhymer.java
@@ -1,16 +1,21 @@
 package edu.kis.vh.nursery;
 
-public class FIFORhymer extends DefaultCountingOutRhymer {
+public class FIFORhymer extends defaultCountingOutRhymer {
 
-	public DefaultCountingOutRhymer temp = new DefaultCountingOutRhymer();
+	public defaultCountingOutRhymer temp = new defaultCountingOutRhymer();
 	
 	@Override
 	public int countOut() {
 		while (!callCheck())
-			temp.countIn(super.countOut());
+			
+		temp.countIn(super.countOut());
+		
 		int ret = temp.countOut();
+		
 		while (!temp.callCheck())
-			countIn(temp.countOut());
+			
+		countIn(temp.countOut());
+		
 		return ret;
 	}
 }
