+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -1,6 +1,6 @@
 package edu.kis.vh.nursery;
 
-public class HanoiRhymer extends DefaultCountingOutRhymer {
+public class HanoiRhymer extends defaultCountingOutRhymer {
 
 int totalRejected = 0;
 
@@ -10,8 +10,8 @@ int totalRejected = 0;
 
 	public void countIn(int in) {
 	if (!callCheck() && in > peekaboo())
-		totalRejected++;
-	else
-		super.countIn(in);
+			totalRejected++;
+			else
+				super.countIn(in);
 	}
 }
