+++ src/main/java/edu/kis/vh/nursery/factory/DefaultRhymersFactory.java
@@ -1,28 +1,29 @@
 package edu.kis.vh.nursery.factory;
 
-import edu.kis.vh.nursery.DefaultCountingOutRhymer;
+import edu.kis.vh.nursery.defaultCountingOutRhymer;
 import edu.kis.vh.nursery.FIFORhymer;
 import edu.kis.vh.nursery.HanoiRhymer;
+import edu.kis.vh.nursery.factory.Rhymersfactory;
 
 public class DefaultRhymersFactory implements Rhymersfactory {
 
 	@Override
-	public DefaultCountingOutRhymer getStandardRhymer() {
-		return new DefaultCountingOutRhymer();
+	public defaultCountingOutRhymer GetStandardRhymer() {
+		return new defaultCountingOutRhymer();
 	}
 
 	@Override
-	public DefaultCountingOutRhymer getFalseRhymer() {
-		return new DefaultCountingOutRhymer();
+	public defaultCountingOutRhymer GetFalseRhymer() {
+		return new defaultCountingOutRhymer();
 	}
 
 	@Override
-	public DefaultCountingOutRhymer getFIFORhymer() {
+	public defaultCountingOutRhymer GetFIFORhymer() {
 		return new FIFORhymer();
 	}
 
 	@Override
-	public DefaultCountingOutRhymer getHanoiRhymer() {
+	public defaultCountingOutRhymer GetHanoiRhymer() {
 		return new HanoiRhymer();
 	}
