+++ src/main/java/edu/kis/vh/nursery/factory/Rhymersfactory.java
@@ -1,15 +1,15 @@
 package edu.kis.vh.nursery.factory;
 
-import edu.kis.vh.nursery.DefaultCountingOutRhymer;
+import edu.kis.vh.nursery.defaultCountingOutRhymer;
 
-	public interface Rhymersfactory {
+public interface Rhymersfactory {
 
-		public DefaultCountingOutRhymer getStandardRhymer();
-		
-		public DefaultCountingOutRhymer getFalseRhymer();
-		
-		public DefaultCountingOutRhymer getFIFORhymer();
-		
-		public DefaultCountingOutRhymer getHanoiRhymer();
-		
-	}
+	public defaultCountingOutRhymer GetStandardRhymer();
+
+	public defaultCountingOutRhymer GetFalseRhymer();
+
+	public defaultCountingOutRhymer GetFIFORhymer();
+
+	public defaultCountingOutRhymer GetHanoiRhymer();
+
+}
