+++ src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -8,5 +8,5 @@ public class Node {
 	public Node(int i) {
 		value = i;
 	}
-	
+
 }
